+++ aikau/src/main/resources/alfresco/forms/controls/BaseFormControl.js
@@ -418,6 +418,18 @@ define(["dojo/_base/declare",
          {
             this.wrappedWidget.set("disabled", status);
          }
+         if (this.domNode)
+         {
+            if (status)
+            {
+               domClass.add(this.domNode, "alfresco-forms-controls-BaseFormControl--disabled");
+            }
+            else
+            {
+               domClass.remove(this.domNode, "alfresco-forms-controls-BaseFormControl--disabled");
+            }
+         }
+         
          this.validate();
       },
 
@@ -1318,6 +1330,13 @@ define(["dojo/_base/declare",
                   evt.preventFocusTheft = true;
                }
             }));
+
+            // See AKU-1049 - emit a custom event to be captured by the form to let it know to re-publish
+            // all the field values. This will allow "late" created fields to process rules...
+            on.emit(this.domNode, "ALF_FIELD_ADDED_TO_FORM", {
+               bubbles: true,
+               cancelable: true
+            });
          }
          else
          {